start pytest for com1DFA
